gjort klart tabort, nu är den säkrare

+++ Wiki/funktioner/tabort.php
@@ -4,8 +4,9 @@
 
 session_start();
 
-include('../../Databas/dbh.inc.php');
-include("../../json/felhantering.php");
+include('./Databas/dbh.inc.php');
+//include("../../json/felhantering.php");
+include("./api_anvandare.php");
 
 if(!empty($_POST['nyckel'])){ // Kollar efter om api-nyckeln är tom
     
@@ -20,10 +21,10 @@ if(!empty($_POST['nyckel'])){ // Kollar efter om api-nyckeln är tom
         switch ($_POST['funktion']) {
 
             case 'tabortWiki':
-                tabortWiki($conn);
+                tabortWiki(getAnvandare($conn),$conn);
                 break;
             case 'tabortWikiSida':
-                tabortWikiSida($conn);
+                tabortWikiSida(getAnvandare($conn),$conn);
                 break;
             
             default:
@@ -38,35 +39,66 @@ if(!empty($_POST['nyckel'])){ // Kollar efter om api-nyckeln är tom
 else {
 hantering('401','Api-nyckeln är inte definerad.');
 }
-function tabortWiki($conn){
+function tabortWiki($anvandarId,$conn){
     //-include('dbh.inc.php');
     $wikiId = $_POST['wikiId'];
-
+    
     $delTjanst = "DELETE FROM tjanst WHERE id='{$wikiId}'";
+    
+    $redan_dolt = $conn->query("SELECT * FROM wiki WHERE id ='{$wikiId}'");
+    $row = $redan_dolt->fetch_assoc();
+    $tjanstId=$row["tjanstId"];
 
     
-    if(mysqli_query($conn, $delTjanst)){
-        hantering('202','tog bort wikin');
-    } else {
-        hantering('400','kunde ej exekvera');
+
+    $tjanst= $conn ->query("SELECT * FROM tjanst WHERE id = '$tjanstId'");
+    $row = $tjanst->fetch_assoc();
+    $anvandarId2=$row["anvandarId"];
+
+    if($anvandarId==$anvandarId2){
+    
+        if(mysqli_query($conn, $delTjanst)){
+            hantering('202','tog bort wikin');
+        } else {
+            hantering('400','kunde ej exekvera');
+        }
+    }
+    else{
+        hantering('400','något har gått fel med anvandare');
     }
 
     $conn->close();
 
 }
 
-function tabortWikiSida($conn){
-    
-
+function tabortWikiSida($anvandarId,$conn){
     $sidID = $_POST["sidId"];
 
+    $wikiSida = $conn->query('select * from wikisidor where id ='.$sidID);
+
+    $row = $wikiSida->fetch_assoc();
+    $wikiId=$row["wikiId"];
+
+    $wiki = $conn->query("SELECT * FROM wiki WHERE id ='{$wikiId}'");
+    $row = $wiki->fetch_assoc();
+    $tjanstId=$row["tjanstId"];
+
+    $tjanst= $conn ->query("SELECT * FROM tjanst WHERE id = '$tjanstId'");
+    $row = $tjanst->fetch_assoc();
+    $anvandarId2=$row["anvandarId"];
+
+    if($anvandarId==$anvandarId2){
     
-    $delSida = "DELETE FROM wikisidor WHERE id = '{$sidID}'";
-    if(mysqli_query($delSida)){
-        hantering('202','tog bort wikisidan');
+        $delSida = "DELETE FROM wikisidor WHERE id = '{$sidID}'";
+        if(mysqli_query($conn,$delSida)){
+            hantering('202','tog bort wikisidan');
+        }
+        else{
+            hantering('400','kunde ej exekvera');
+        }
     }
     else{
-        hantering('400','kunde ej exekvera');
+        hantering('400','något har gått fel med anvandare');
     }
 }
